Solar system sim

+++ src/core/ecs/SystemManager.ts
@@ -1,21 +1,23 @@
-import { System } from "./System";
-import { World } from "./World";
+import { ISystem } from "./ISystem";
+import { IWorld } from "./IWorld";
+import { IEventEmitter } from "../events/IEventEmitter";
 import { EventEmitter } from "../events/EventEmitter";
+import { ISystemManager } from "./ISystemManager";
 
 /**
  * Manages all systems in the ECS architecture.
  * Responsible for registering, updating, and retrieving systems.
  */
-export class SystemManager {
+export class SystemManager implements ISystemManager {
   /**
    * The collection of registered systems.
    */
-  private systems: System[] = [];
+  private systems: ISystem[] = [];
 
   /**
    * Event emitter for system-related events.
    */
-  private eventEmitter: EventEmitter = new EventEmitter();
+  private eventEmitter: IEventEmitter = new EventEmitter();
 
   /**
    * Registers a system with the manager.
@@ -24,11 +26,11 @@ export class SystemManager {
    * @param system The system to register
    * @param world The world the system is being registered with
    */
-  public registerSystem(system: System, world?: World): void {
+  public registerSystem(system: ISystem, world?: IWorld): void {
     this.systems.push(system);
 
     // Call the system's onRegister method if a world is provided
-    if (world) {
+    if (world && system.onRegister) {
       system.onRegister(world);
     }
 
@@ -40,7 +42,7 @@ export class SystemManager {
    *
    * @param callback The callback function
    */
-  public onSystemRegistered(callback: (system: System) => void): void {
+  public onSystemRegistered(callback: (system: ISystem) => void): void {
     this.eventEmitter.on("systemRegistered", callback);
   }
 
@@ -50,7 +52,8 @@ export class SystemManager {
    * @param world The world to update
    * @param deltaTime The time elapsed since the last update
    */
-  public updateAll(world: World, deltaTime: number): void {
+  public updateAll(world: IWorld, deltaTime: number): void {
+    this.systems.sort((a, b) => a.priority - b.priority);
     this.systems.forEach((system) => {
       system.update(world, deltaTime);
     });
@@ -62,10 +65,10 @@ export class SystemManager {
    * @param systemType The constructor of the system type
    * @returns The system instance, or undefined if not found
    */
-  public getSystem<T extends System>(
+  public getSystem<T extends ISystem>(
     systemType: new (...args: any[]) => T
   ): T | undefined {
-    return this.systems.find((system) => system instanceof systemType) as T;
+    return this.systems.find((system) => system instanceof (systemType as any)) as T;
   }
 
   /**
@@ -76,11 +79,11 @@ export class SystemManager {
    * @param world The world the system is being removed from
    * @returns True if the system was removed, false if it wasn't found
    */
-  public removeSystem(system: System, world?: World): boolean {
+  public removeSystem(system: ISystem, world?: IWorld): boolean {
     const index = this.systems.indexOf(system);
     if (index !== -1) {
       // Call the system's onRemove method if a world is provided
-      if (world) {
+      if (world && system.onRemove) {
         system.onRemove(world);
       }
 
@@ -96,15 +99,26 @@ export class SystemManager {
    *
    * @param world The world the systems are being removed from
    */
-  public clear(world?: World): void {
+  public clear(world?: IWorld): void {
     if (world) {
       // Call onRemove for each system
       this.systems.forEach((system) => {
-        system.onRemove(world);
+        if (system.onRemove) {
+          system.onRemove(world);
+        }
       });
     }
 
     this.systems = [];
+  }
+
+  /**
+   * Gets all registered systems.
+   *
+   * @returns An array of all registered system instances
+   */
+  public getAllSystems(): ISystem[] {
+    return this.systems;
   }
 
   /**
@@ -113,7 +127,7 @@ export class SystemManager {
    * @param system The system that was registered
    * @private
    */
-  private emitSystemRegistered(system: System): void {
+  private emitSystemRegistered(system: ISystem): void {
     this.eventEmitter.emit("systemRegistered", system);
   }
 }
